+++ components/Charts/ChartConfigurator.vue
